requirements.txt
black==21.6b0
Django==3.2.4
django-debug-toolbar==3.2.1
Faker==8.8.0
psycopg2-binary==2.8.6
